Use plugin bom version prior to trilead API upgrade
The plugin compatibility tester mistakenly uses the version of optional
plugins that is provided by the pom of the tested plugin rather than
using the version of plugin that is listed in the pom file of the plugin
bill of materials.

Because of that issue, the 2.426.x plugin bill of materials version
needs to remain on an older version in order to not include an optional
dependency on a newer version of the trilead API plugin.

3080.vfa_b_e4a_a_39b_44 is the version immediately before a trilean API
plugin upgrade in the BOM.

https://github.com/jenkinsci/bom/pull/3404 has more details

https://github.com/jenkinsci/ssh-credentials-plugin/pull/211 includes
some discussion as well.